Change update checker action to directly update Cody plugin (#7545)
## Changes

This potentially my fix one of the most reported exceptions:
https://sourcegraph.sentry.io/issues/6396868697/

Either way I changed the action to update the Cody plugin automatically,
instead of just moving user to the plugins list.

I also investigated if we can reload our plugin without restarts.
Currently it does not work, and it would require some fixes to
disposables and services we have, but in principle it should be
possible. If someone would be looking at this later, some useful
resources:

Docs:

https://plugins.jetbrains.com/docs/intellij/dynamic-plugins.html?from=jetbrains.org#restrictions

Debugging issues:

https://intellij-support.jetbrains.com/hc/en-us/community/posts/360008211319-Plugin-not-unload-safe-because-class-loader-cannot-be-unloaded

## Test plan

1. Run IntelliJ, you should see new message from update checker:


![image](https://github.com/user-attachments/assets/e9a0889e-50fb-41f7-b9dc-cf97365408f3)

2. Click update
3. Wait a dozen of seconds, and popup with notification about required
restart should appear:


![image](https://github.com/user-attachments/assets/b3daff06-76ce-4120-86ae-7e04099687a9)