feat(auto-edit): create a new fireworks-websocket adapter for auto-edit to use WebSocket instead of HTTP for connections, #CODY-5334 (#7490)
Use WebSocket to keep a long lived connection to the server to avoid
cross-request connection establishment overhead. Connections are
established on first send operation.

Currently, this feature requires a WebSocket to HTTP proxy to work. The
new feature requires following changes to user settings:

```
    "cody.experimental.autoedit.config.override": {
        "provider": "fireworks-websocket",
        "webSocketEndpoint": "ws://0.0.0.0:3000",
```

This is a highly experimental feature, guarded by
`auto-edit-use-web-socket-for-connections` feature flag.

## Test plan

- Follow instructions from
[here](https://www.notion.so/sourcegraph/Using-WebSocket-For-Cody-Model-Requests-1bda8e112658802ca85cd73ba60f98dc?showMoveTo=true&saveParent=true)
- internal testing with feature flag

---------

Co-authored-by: Valery Bugakov <skymk1@gmail.com>