add index to props
The `HumanMessageEditor` component now receives an `index` prop, which is used to determine if it is the last editor in the list. The editor state is only reset if the `index` prop is equal to the index of the last editor. This ensures that only the last editor is reset when a new message is added to the transcript.