feat(docker): The docker container for cody-cli
Finally got a fully working, locally built docker container that does
not depend on `npm`. Had to add a .dockerignore so that it won't include
`node_modules` when building, and also removed `.git` files.

I only tested this manually by running `docker run --rm cody-cli cody
help` which does display the help files correctly. There's some built in
testing when creating the docker container (`which cody`, etc) so we
know it doesn't build broken images.

I decided against using `pnpm link -g` since it requires a bunch of
setup and configuration, and instead created a shell script that calls
node on the built js files, and added that to the path