feat(audoedit): add basic client cache (#7541)
- Implements basic client caching, which is not yet useful for latency.
I started working on the in-flight requests reuse on the same branch and
then realized that it's better to split them so that we can iterate on
the request manager together with @umpox in parallel.
- I'll create a separate PR for the in-flight requests recycling.
One awkward thing this PR does is add a request manager to the renderer
manager constructor so that we can prune cached suggestions on
acceptance or rejection. In a follow-up, we can think of a cleaner way
of doing that.