fix(modes): Hide agentic mode when edit is not enabled (#7555)
This commit updates the `ModeSelectorButton` component to hide the
agentic mode option when the edit functionality is not enabled. This
ensures that users do not see the agentic mode option if they cannot use
it.

This includes Web client


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Agent mode should be hidden for clients that dont support edit