Pass full filepath in prompt to the LLM.  (#7472)
closes:
https://linear.app/sourcegraph/issue/CODY-5393/bug-fix-send-the-file-path-to-the-model-for-at-mentions

This PR updates the get text from editor state logic to include the full
file display path and not just the file basename.

The text is sent to the LLM as the human message text.

## Test plan

unit tests added. 

mention a file and debug & test if the full file display path is
included in the prompt.