[JetBrains] Run agent once per test class (#7553)
We had a major issue in our integration tests. `setUp` and `tearDown`
methods are called before and after EACH case. That means we initialized
(and shutdown-ed) the agent not once per the class but multiple times.
And the recordings were recorded multiple times overriding the previous
version. We were lucky it worked for document code test class.

This PR is a prerequisite for integration tests for autocomplete.

Additionally, this PR provides a mechanism to specify the recording name
that will help us later with autocomplete and probably other test
classes.

It will likely speed up tests. Let's see once CI passes.

## Test plan
- green ci
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->