[JetBrains] Switch to new autocompletions API (#7456)
<!-- start git-machete generated -->

This PR switches to the new autocompletions API. 

There is a single significant change - before JB client based on an
assumption that the text to be inserted starts exactly at the cursor
position. New API tends to send whole lines instead. We got ranges
anyway so it seems to be more reasonable to use them to derive the
position. Cursor is a moving part. Switching to ranges looks like an
improvement for me.

The other thing is that for RD (Remote Development) we use a separate
mechanism (completion providers). There was a bug there that appeared
now, as the completion texts starts at the beginning of the line. Fixed.

# Based on PR #7311

## Chain of upstream PRs as of 2025-03-18

* PR #7311:
  `main` ← `tr/autoedit-agent-support`

  * **PR #7456 (THIS ONE)**:
    `tr/autoedit-agent-support` ← `mkondratek/chore/new-completions-api`

<!-- end git-machete generated -->

## Test plan
Cody autocomplete works.

You can try this piece of code:
```kt
fun main() {
    val list = listOf(1, 2, 3, 4, 5)
    for (<A>) {
        println(x)
        // todo: calculate <B>
    }
}
```
Triggered at `<A>` should show `x in list`.
Triggered at `<B>` should show `the sum of the list`. 

The exact results can be different.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->