fix/guardrails: In enforced mode, hide smart apply, etc. actions until attribution complete (#7579)
When Guardrails is in "enforced" mode, we want to prevent the user from
using code until an attribution check is complete. Hide the action
buttons until it is OK to show the code.

For Guardrails "permissive" mode, or no Guardrails, `showCode` is always
`true` and we will show the action buttons as soon as the code is
complete.

This also bumps the timeout for Guardrails attribution requests. These
requests can easily take 20s+ for long code blocks.

Fixes CODY-5552, CODY-5551

## Test plan

Tested manually:

1. Set up an instance with `"attribution.enabled": true,
"attribution.mode": "enforced"` and connect to it.
2. Generate more than 10 lines of code.
3. Check that the action buttons do not appear while "Checking
Guardrails"