refactor(chat): Remove loading dots from context display (#7557)
Follows up on https://github.com/sourcegraph/cody/pull/7502 which add
the LoadingDots at Transcript level, where the LoadingDots will show up
at the end of the transcript when there is an assistant message in
progress.

This commit refactors the `ContextCell` component to remove the
`LoadingDots` component and simplify the display of context items.

The `LoadingDots` component was previously used to indicate that the
context was loading. However, it was only displayed when
`isContextLoading` was true and `isAgenticChat` was false. This logic
was removed, and the context items are now always displayed within the
`AccordionContent`.

This change simplifies the component and improves the user experience by
providing a consistent display of context items, regardless of the
loading state or chat mode.

## Test Plan

Ensures there are no multple loading dots showing up in the UI at any
time.

(cherry picked from commit 656e2c3e8537b43c066885a23ac3afec79f60065)