fix/chat: Label the button that runs shell commands "Execute" instead of "Execute in Terminal" (#7594)
Also move the ... overflow menu to be the last menu item on the left.

## Test plan

Tested manually: Ask Cody to generate a 10 line shell script, and verify
that the button is labelled "Execute."

Check that the Copy and Execute icons are aligned, both in cases where
there are filenames/Guardrails, and where there is no metadata.

![Screenshot 2025-03-28 at 16 55
11](https://github.com/user-attachments/assets/171ca72a-2bf6-4fda-98be-7ec7da0d7235)

(cherry picked from commit 218166c9933fe8d2eaa77249781743a992d654dc)