fix(autoedit): Fix case where no completion provider would be registered for non VS Code clients (#7640)
## Description

Fixes an issue where we would never register the default completion
provider for clients where autoedit was enabled as a client capability.
This was because we assumed if the client enabled autoedits then all
users would get it. That's not the case as it depends on user
eligibility.

Now the logic should follow the same as VS Code.

## Test plan

1. Testing with integration tests that mock network recordings

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->