[JetBrains] Add autoedit support (#7455)
Resolves https://linear.app/sourcegraph/issue/CODY-5287.

---

For now, this should be considered **a Beta version**—not everything has
to work perfectly or be fully optimized.
Two specific things to keep in mind:
- The **scope** in which auto-edit suggests changes—it would be great if
it covered entire blocks (methods, if statements, for and while loops,
etc.). This might not be as noticeable in real-world use, but during
testing, it stands out when, for example, one more line could have been
included but wasn’t.
- **It doesn't work on Windows** (integration tests pass, but there’s an
issue at the UI level). It’s strange because I saw it working fine just
yesterday. I’m getting an error and will be working on a fix, but it
might come in future PRs.

<!-- start git-machete generated -->

# Based on PR #7311

## Chain of upstream PRs as of 2025-03-18

* PR #7311:
  `main` ← `tr/autoedit-agent-support`

  * **PR #7456:
  `tr/autoedit-agent-support` ← `mkondratek/chore/new-completions-api`

    * **PR #7455 (THIS ONE)**:
`mkondratek/chore/new-completions-api` ←
`mkondratek/feat/autoedits-diff-v2`


<!-- end git-machete generated -->

---

This PR brings Autoedits into JetBrains.

Autoedits are still in an experimental phase. To turn them on please add
the following entry to your `cody_settings.json`:
```
  "cody.suggestions.mode": "auto-edit (Beta)",
```

I will provide integration tests later (work in progress):
https://github.com/sourcegraph/cody/pull/7584

## Test plan

Feel free to use this repo and dir as a playground and source of
examples:

https://github.com/sourcegraph/cody-chat-eval/tree/main/src/autoedit_debug/examples/renderer-testing-examples

---
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->