Crash occurred when user tries to save new code file (#7645)
`VfsUtil.toUri()` interpert the drive letter in the windows path as a
scheme in the uri. This leads to a URI object which, when the `toPath()`
method is called, causes an error. This prevents the creation of a new
code file and spoils all "Save file" functionality. This PR tries to
remedy the situation by conditionally adding a "file" prefix to the uri
string. Using `withScheme()` does not fix the situation because the
scheme is already mistakenly set when the URI is created.

## Test plan
1. Use chat to generate some code
2. Under code block click on Save file icon
3. Add name for file
4. Click on 'OK'
5. Code file should be created and opened in the editor without errors
<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Tomasz Gołębiowski <tgolebiowski@virtuslab.com>