feat(agent): Manually cleanup listeners from the socket (#7669)
## Changes

Looks like in case of the socket we cannot rely on the automatic cleanup
upon the object destruction, because we are reusing the sockets. That is
why we need to manually remove listeners when the socket is closed.

## Test plan

1. Play with Cody for a while (1 min of editing a text file is more than
enough)
2. Search log for "leak detected" - there should be no leaks or too many
listener warnings.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

(cherry picked from commit 49cdd04ecaecdf3cbb50f272ef8fbe945c137996)