feat(auto-edit): reuse inflight requests (#7661)
- Enables in-flight request reuse for the same line or next added new
line if, while we were waiting for the request, the user only inserted
text. If more than one line is added or modified or any test is deleted,
we do not attempt to reuse the received response for the in-flight
request.
- Built on top of https://github.com/sourcegraph/cody/pull/7590. I fixed
a couple of unit tests, @umpox, related to the streaming changes. Let's
attempt to land your streaming PR, and I'll work on resolving the merge
conflicts.