Pass sourcegraph instance's endpoint to websocket for authentication purposes
SAMS library only supports machine-to-machine authentication of services. As a result, the token generated from sourcegraph instance does not apply. To workaround this, we now send both the endpoint and token to websocket service, which will make a HTTP call to the instance to re-verify before upgrading the connection to websocket.