chore(core): Fix unhandled AbortError errors  (#7714)
## Changes

`GraphQLResultCache::get` is used by `getSiteVersion`, and it by turn is
user to create an observable.
Code ran in observables cannot throw exceptions.
Signature of the `get` was suggesting it returns `Errors` by value
instead of throwing them, but both options were possible.
Error handling which I aeed makes sure errors are always retuned by
value.

## Test plan

1. Run `pnpm -C vscode run test:integration` few times locally.
2. You should see no `AbortError` errors.

It was reproducing better on M84 for me, but on main error is still
visible in the log although it does not crash the test.
With this PR error is totally gone, and on M84 tests is not crashing
anymore.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->