feat(autoedit): Implement hot streak (#7652)
This PR implements:
- Hot streak suggestions
- Next cursor predictions
- Fuzzy caching

### Hot streak

We can now stream much longer responses without worrying about latency.
We will emit the suggestion after 5 lines, assuming we have a diff.
Then, future suggestions are cached using the same `hotStreakId`. Once
you accept one suggestion you can then move to the next one, view that
instantly, accept and so on...

In order to reliably produce suggestions whilst the LLM is still
streaming, we look for the first "unchanged" hunk in the remaining
prediction, where there is a change before this hunk. This unchanged
area gives us a stable point where we can reliably produce a diff and
`codeToRewrite` area, without worrying about handling edge cases like
lines being added before/after the range.

### Next cursor predictions

Each suggested auto-edit response now includes an `editPosition`. This
is essentially the first changed line of this particular suggestion. For
hot-streak this allows us to support next cursor suggestions. I.e., you
accept the first suggestion, we look in the cache for a matching item
with the same `hotStreakId`. If we find it, we move the cursor there.

In a [future PR](https://github.com/sourcegraph/cody/pull/7701) we will
implement a UI to show these suggestions instead of moving the cursor
automatically. We will also need a small bit of logic to decide if we
should show the prediction immediately or if we should show the next
cursor UI.

I have also included `editPosition` for non hot-streak responses. This
will allow us to handle cases where you might get a long response with
lots of unchanged lines and a final change near the end. This would not
typically produce a hot-streak, but we'll likely want the next cursor UI
here as the suggestion will be far away from the users' cursor position.
Always providing `editPosition` also means we can do some more complex
stuff around next cursor predictions in the future.

### Fuzzy caching

This PR introduces some level of fuzzy caching in order to make sure the
hot-streak response chaining works. Essentially we now have a single
strict + unique `cacheId` as the cache key for each item. This is just a
random uuid and only serves as a way to reliably eject from the cache
when needed. Most of the cache logic will be done fuzzily instead.

We look for valid suggestions where the rewrite window around the
suggestion is still present in the `document`. We then retrieve the item
that is closest to the cursor position. The cache doesn't take into
account things like edit history or other context items yet. We may want
to revisit this soon. For the purposes of hot-streak, we also take into
account the last accepted `hotStreakId`, if this is present we will
filter cache items to match against this id.

---------

Co-authored-by: Valery Bugakov <skymk1@gmail.com>