fix(smart-apply): error when applying code to an empty file (#7439)
Refs CODY-5396 
Fixed bug where inserting any code within an empty file triggered an
error
which doesn't abort the smart apply action, leaving the user to think
it's super slow.

This started after the work here
https://github.com/sourcegraph/cody/pull/6768
Because `startStreamingEditTask` calls fixup `startDecorator` 

https://github.com/sourcegraph/cody/pull/6768/files#diff-aab0c624f666dd0c22f4415b5d9cf6e2b1a9d3ceb14526e2714896beade885dfR291

And we bump the end line by one even when the file is empty

https://github.com/sourcegraph/cody/blob/main/vscode/src/non-stop/FixupTask.ts#L155
The `trimEmptyLinesFromRange` ended up with an index higher than
available which
crashed
https://github.com/sourcegraph/cody/pull/7439/files#diff-e0078048d45c5a1b5f5eea5af56c3919ecf75b0fe610dfee87578e87d0944ae0R87

Also used the correct model name to ensure we get 15k input instead of
the default
fallback of 7k in case we have a large file that requires edits.

## Test plan
Apply code in an empty file and observe it working

---------

Co-authored-by: Beatrix <68532117+abeatrix@users.noreply.github.com>