fix(search): handle missing keywords in XML parsing (#7777)
CLOSE
https://linear.app/sourcegraph/issue/CODY-5475/did-you-mean-doesnt-work-in-chat-mode

This commit addresses a potential issue where the `document.keywords`
property might be missing or undefined during XML parsing. It adds a
check to ensure that `document?.keywords` exists before attempting to
access its properties. This prevents potential errors and ensures that
the code gracefully handles cases where the XML structure does not
contain the expected keywords.

## Test Plan

1. Open the mode selector, and select Chat as the mode
2. Ask Cody 'where is auth?'
3. No error shows up

Dot com:


![image](https://github.com/user-attachments/assets/12da2f01-993c-4cad-99aa-dee441f60f6d)

S2:


![image](https://github.com/user-attachments/assets/65cba3ea-337d-40f6-a8b0-4b45bfa866ba)


### Before


![image](https://github.com/user-attachments/assets/3adf6e8c-b767-4d03-bfe8-39cdf60b9fdf)

## Note

Clicking on the "Did you mean" will not execute the search query due to
an existing bug with the mode selector, which will be fixed in
https://github.com/sourcegraph/cody/pull/7776