Fix search results not getting mentioned in follow up chat (#7711)
closes:
https://linear.app/sourcegraph/issue/CODY-5680/on-non-pro-non-free-plan-vs-code-latest-pre-release-after-triggering-a

In search mode, once we get the results, we add a results mention chip
to the last editor ref.

As part of https://github.com/sourcegraph/cody/pull/7550/files, the
condition to pass the lastEditorRef to the last TranscriptInteraction
was changed and made agent mode specific. But that logic doesn't work
for the other modes and the lastEditorRef was always null.

This PR fix that logic so that the last editor ref is properly set. This
PR also removes `onAddToFollowupChat` which is no longer used now.

## Test plan

- make an interaction using search mode
- it should add a search results mention chip in the follow up chat
- selecting/deselecting the search results using the checkbox should
update the mention chip

<img width="822" alt="image"
src="https://github.com/user-attachments/assets/d2f888a4-0d1e-4b38-968a-9af6666f0b10"
/>