Make sure autocomplete service is working per project (#7785)
## Changes

`CodyAutocompleteManager` should not be a global service, but
project-level service.
The same applies to utility methods like a
`CodyEditorUtil.getAllOpenEditors`.
Until now it was implemented in a bit messy way and there was a risk of
accidental running actions for files which were coming from other
projects.
That was happening e.g. in the `registerAllOpenedFiles` during the
startup.
I have so proof it was causing any particular issue (except causing
additional memory and CPU usage) but there is some chance it could
contribute to or cause issues like this:
https://linear.app/sourcegraph/issue/BUGS-1580/bug-jetbrains-rider-freezes-when-switching-files-and-on-startup

## Test plan

Full QA + testing scenarios from `testing.md`, `Autocomplete` section.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->