fix(auto-edit): model call latency metric did not reflect actual timing (#7783)
`getAndProcessModelResponses` returns an `AsyncGenerator`. As a result,
the wait does not happen until calls to `next` function. This causes
timing skewed to only function call latencies. The fix is to record
latency at the time `next` function is called instead.

ref: CODY-5528

## Test plan

verified by manually trying autoedit in debug environment and verify the
metrics in metric explorer.
![Screenshot 2025-04-22 at 16 10
27](https://github.com/user-attachments/assets/43936eb7-1701-4479-8deb-66d021610b47)