chore(auto-edit): prepare rollout of websocket proxy to s2 users (#7781)
Reduce friction of rolling out websocket proxy for s2 users by

- force using fireworks-websocket adapter if user has websocket feature
flag enabled and user is an S2 user
- remove the need for users to override their settings in order to use
websocket proxy. For example, `apiKey` was required previously, but now
given it's stored on the proxy, user shouldn't need to specify the key
anymore.
- remove the model parameter the websocket proxy currently use direct
connect to fireworks, which does not have the option of specifying
models (if wrong model name is specified, user may get errors).

With these, S2 users shouldn't need to make any settings change in order
to use the new proxy for auto-completion.

## Test plan

- local manual testing