Don't execute Cody calls on EDT (#7792)
Probably fixes
https://linear.app/sourcegraph/issue/BUGS-1580/bug-jetbrains-rider-freezes-when-switching-files-and-on-startup

## Changes
Previously `codyAgent.server.textDocument_didOpen` and
`codyAgent.server.textDocument_didFocus` were called on the EDT with the
rest of the code for obtaining document content. That is bad, because if
for any reason Cody will be stuck or slow it will make whole IntelliJ
unstable. Cody calls should always happen on separate background thread.

## Test plan

1. Open a lot of big files
2. Restart IntelliJ
3. There should be no hangs of UI when starting Cody (compared to doing
the same with Cody disabled/uninstalled)

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->